bug fixes for meshes make __module__ attribute set for the intended destination module for classes
