+++ discretize/cylindrical_mesh.py
@@ -1729,6 +1729,9 @@ class CylindricalMesh(
     )
 
 
+CylindricalMesh.__module__ = 'discretize'
+
+
 @deprecate_class(removal_version="1.0.0")
 class CylMesh(CylindricalMesh):
     pass
