+++ discretize/operators/inner_products.py
@@ -864,3 +864,6 @@ class InnerProducts(object):
         "getEdgeInnerProductDeriv",
         removal_version="1.0.0",
     )
+
+
+InnerProducts.__module__ = 'discretize.operators'
